PUBLISH: v1.0.5

+++ Test/Apps72.Dev.Data.Tests/Entities.cs
@@ -42,7 +42,7 @@ namespace Data.Tests.Entities
         /// <summary />
         public virtual DateTime? HIREDATE { get; set; }
         /// <summary />
-        public virtual Int32? SAL { get; set; }
+        public virtual Decimal? SAL { get; set; }
         /// <summary />
         public virtual Int32? COMM { get; set; }
         /// <summary />
